convert simmer final_equity_spread and compute_win_prob to i32
final_equity_spread was computing an i32 spread (scores and leave
values are all i32 millipoints) and pointlessly casting to f32.
Return i32 directly. compute_win_prob takes i32, uses signum()
for the game-ended case, and casts to f64 only for the sigmoid.
Callers in move_picker and main_simmer no longer need f32 casts
for sim_spread subtraction.

+++ src/main_simmer.rs
@@ -314,7 +314,7 @@ fn main() -> error::Returns<()> {
             );
             let final_spread = simmer.simmer.final_equity_spread();
             let win_prob = simmer.simmer.compute_win_prob(game_ended, final_spread);
-            let sim_spread = final_spread - simmer.simmer.initial_score_spread as f32;
+            let sim_spread = final_spread - simmer.simmer.initial_score_spread;
             candidate
                 .stats
                 .update(sim_spread as f64 + win_prob * simmer.simmer.win_prob_weightage());
