+++ src/move_picker.rs
@@ -162,7 +162,7 @@ impl<N: kwg::Node, L: kwg::Node> MovePicker<'_, N, L> {
                         );
                         let final_spread = simmer.simmer.final_equity_spread();
                         let win_prob = simmer.simmer.compute_win_prob(game_ended, final_spread);
-                        let sim_spread = final_spread - simmer.simmer.initial_score_spread as f32;
+                        let sim_spread = final_spread - simmer.simmer.initial_score_spread;
                         candidate.stats.update(
                             sim_spread as f64 + win_prob * simmer.simmer.win_prob_weightage(),
                         );
